requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.1
altair==5.0.1
appdirs==1.4.4
async-timeout==4.0.2
brotlipy==0.7.0
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cuda-python==12.1.0
cycler==0.11.0
Cython==0.29.35
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
fastapi==0.97.0
ffmpy==0.3.0
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
gradio==3.35.2
gradio_client==0.2.7
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
imageio==2.31.1
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
lightning-utilities==0.8.0
linkify-it-py==2.0.2
markdown-it-py==2.2.0
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl-fft==1.3.6
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.4
munch==3.0.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
orjson==3.9.1
pandas==2.0.2
pathtools==0.1.2
Pillow==9.4.0
ply==3.11
pretrainedmodels==0.7.4
protobuf==4.23.3
pydantic==1.10.9
pydub==0.25.1
pyparsing==3.0.9
PyQt5==5.15.7
python-multipart==0.0.6
pytorch-lightning==2.0.3
PyWavelets==1.4.1
pywin32==305.1
qudida==0.0.4
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
segmentation-models-pytorch==0.3.3
semantic-version==2.10.0
sentry-sdk==1.25.1
setproctitle==1.3.2
smmap==5.0.0
starlette==0.27.0
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
toolz==0.12.0
torch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.0
tqdm==4.65.0
tzdata==2023.3
uc-micro-py==1.0.2
uvicorn==0.22.0
wandb==0.15.4
webencodings==0.5.1
websockets==11.0.3
yarl==1.9.2